refactor: 폴더 경로 변경
- 파일명 변경
- 클래스명 변경
