+++ src/pages/Search/components/RelatedItem/RelatedItem.module.scss
@@ -1,0 +1,109 @@
+@use "src/utils/styles/mediaQuery.scss" as media;
+
+.suggestion {
+  display: flex;
+  flex-direction: column;
+  position: absolute;
+  width: 100%;
+  overflow: scroll;
+  padding-top: 24px;
+  background-color: white;
+  border-radius: 24px;
+  box-shadow: 2px 3px 12px 1px rgba(0 0 0 / 10%);
+  padding-bottom: 15px;
+  top: -40px;
+  left: 15px;
+  z-index: 1;
+
+  &--hidden {
+    display: none;
+  }
+
+  @include media.media-breakpoint-down(mobile) {
+    width: 93%;
+    display: flex;
+    flex-direction: column;
+  }
+
+  &::-webkit-scrollbar {
+    display: none;
+  }
+
+  &__wrapper {
+    text-decoration-line: none;
+    color: black;
+    border: none;
+    appearance: none;
+  }
+
+  &__item {
+    display: flex;
+    font-size: 20px;
+    height: 48px;
+    padding-left: 40px;
+    padding-right: 57.14px;
+    align-items: center;
+    cursor: pointer;
+
+    &:hover {
+      background: #eeeeee;
+      border-radius: 24px;
+      color: #ff7f23;
+    }
+
+    @include media.media-breakpoint-down(mobile) {
+      font-size: 14px;
+      padding-left: 12px;
+      height: 41px;
+    }
+  }
+
+  &__title {
+    display: flex;
+  }
+
+  &__icon {
+    display: flex;
+    margin-right: 16px;
+    align-items: center;
+
+    @include media.media-breakpoint-down(mobile) {
+      width: 16px;
+      height: 16px;
+    }
+  }
+
+  &__auto-controller {
+    display: flex;
+    flex-direction: row-reverse;
+    margin-top: 25px;
+    align-items: center;
+  }
+
+  &__auto-button-title {
+    display: flex;
+    font-size: 20px;
+    color: #ff7f23;
+    padding: 24px 40px 18px 13px;
+
+    &.active {
+      color: #c4c4c4;
+    }
+  }
+
+  &__toggle-button {
+    display: flex;
+    padding-top: 8px;
+  }
+
+  &__text--not-found {
+    word-break: keep-all;
+    width: 275px;
+    height: 40px;
+    font-size: 16px;
+    line-height: 20px;
+    text-align: center;
+    margin: 0 auto;
+    margin-top: 164px;
+  }
+}
