+++ src/pages/Search/components/ToggleButton/ToggleButton.module.scss
