+++ src/pages/Search/components/ToggleButton/ToggleButton.tsx
@@ -7,7 +7,7 @@ interface Props {
 
 export default function ToggleButton({ onClick, isActive }: Props) {
   return (
-    <button type="button" aria-label="자동완선 on/off" onClick={onClick} data-toggle={isActive ? 'true' : 'false'} className={`${styles.ToggleButton} ${isActive ? styles.active : ''}`}>
+    <button type="button" aria-label="자동완성 토글" onClick={onClick} data-toggle={isActive ? 'true' : 'false'} className={`${styles.ToggleButton} ${isActive ? styles.active : ''}`}>
       <div data-toggle={isActive ? 'true' : 'false'} className={styles.circle} />
     </button>
   );
